+++ R/PomaImpute.R
@@ -35,10 +35,12 @@ PomaImpute <- function(data,
   if (!(method %in% c("none", "half_min", "median", "mean", "min", "knn", "random_forest"))) {
     stop("Incorrect value for method argument")
   }
-  if (missing(method)) {
-    message("method argument is empty. KNN will be used")
-  }
+  # if (missing(method)) {
+  #   message("method argument is empty. KNN will be used")
+  # }
 
+  n_features_raw <- length(rownames(data))
+  
   to_impute <- t(SummarizedExperiment::assay(data)) %>% 
     as.data.frame()
   
@@ -107,9 +109,11 @@ PomaImpute <- function(data,
   }
 
   else if (method == "knn"){
-    imputed_t <- t(to_impute)
-    imputed_res <- impute::impute.knn(imputed_t)
-    imputed <- t(imputed_res$data)
+    suppressWarnings({
+      imputed_t <- t(to_impute)
+      imputed_res <- impute::impute.knn(imputed_t)
+      imputed <- t(imputed_res$data)
+    })
   }
   
   else if (method == "random_forest"){
@@ -128,7 +132,11 @@ PomaImpute <- function(data,
   } else {
     data <- SummarizedExperiment::SummarizedExperiment(assays = t(imputed))
   }
-    
+  
+  n_features_imputed <- length(rownames(data))
+  
+  message(paste0(n_features_raw - n_features_imputed, " features removed."))
+  
   if (validObject(data))
     return(data)
 }